requirements.txt
discord.py
datetime
asyncio
aiohttp
websockets
chadset